+++ src/wazuh_modules/wm_task_general.c
@@ -19,7 +19,6 @@ const char *task_manager_json_keys[] = {
     [WM_TASK_COMMAND] = "command",
     [WM_TASK_PARAMETERS] = "parameters",
     [WM_TASK_AGENTS] = "agents",
-    [WM_TASK_TASKS] = "tasks",
     // Response
     [WM_TASK_ERROR] = "error",
     [WM_TASK_DATA] = "data",
@@ -30,7 +29,11 @@ const char *task_manager_json_keys[] = {
     [WM_TASK_STATUS] = "status",
     [WM_TASK_ERROR_MSG] = "error_msg",
     [WM_TASK_CREATE_TIME] = "create_time",
-    [WM_TASK_LAST_UPDATE_TIME] = "update_time"
+    [WM_TASK_LAST_UPDATE_TIME] = "update_time",
+    // Clean tasks request
+    [WM_TASK_NOW] = "now",
+    [WM_TASK_INTERVAL] = "interval",
+    [WM_TASK_TIMESTAMP] = "timestamp"
 };
 
 const char *task_manager_commands_list[] = {
@@ -40,7 +43,8 @@ const char *task_manager_commands_list[] = {
     [WM_TASK_UPGRADE_UPDATE_STATUS] = "upgrade_update_status",
     [WM_TASK_UPGRADE_RESULT] = "upgrade_result",
     [WM_TASK_UPGRADE_CANCEL_TASKS] = "upgrade_cancel_tasks",
-    [WM_TASK_TASK_RESULT] = "task_result"
+    [WM_TASK_SET_TIMEOUT] = "set_timeout",
+    [WM_TASK_DELETE_OLD] = "delete_old"
 };
 
 const char *task_manager_modules_list[] = {